Create utils.py for misc utility functions, add CreateMenuItem to utils
Call this like...
menu=wx.Menu()
utils.CreateMenuItem(menu, 'some item', some_item_function) # Creates a menu item with the specified label (you can optionally pass position), and binds it to the specified function. Then adds that item to the specified menu.
The order in which you add items to any menu in this manner is importand unless you specify an explicit position, as by default, items get appended to the end of the menu you pass in.