Update import in urban dictionary plugin
**please note**
Currently, the urban dictionary plugin won't work because it imports a file in the same directory as the plugin itself, and sys.path doesn't have that in the list, so imports that aren't in the main queriet app directory or available on the path elsewhere will fail.
I have filed a pull request with Yapsy that adds a callback_after argument to the loadPlugins function.
Once this is merged, we can use the callbacks to get the location of the plugin's directory, and add it to sys.path, let yapsy do it's thing by loading the plugin, then delete said item from the list.
Hacky and rough, but it would allow each plugin to import normally and not worry about this shit.
This bug doesn't affect plugins that use packages elsewhere on the path, but eventually all plugins will come bundled with their dependencies so this needs to work by then.