Update defaultPanels with 'InputPanelWithRadiobox' class
This is the standard input panel, with a text field and search button, just with a radiobox between the 2. It supports all the arguments that 'StandardInputPanel' does, plus choices for the radiobox and an OnRadioboxChange function args.
This should allow for plugins to define more than one type of item to search for.
For example, a dictionary plugin could have definition, synonyms, antonyms, etc, and choose which to search for and display with this radiobox.
You can also pass in a function to run when the radiobox's selection changes, but this is optional.