Work on the keybindings dialog a bit
add docstrings, make methodnames more self explanatory, remove the "save config" button so it's in line with the other panels.
Now, whenever a checkbox or the text field is changed, it calls on_binding_changed (the old on_brb thing), which updates the corresponding entry in the config file for the selected binding. This, of course, doesn't touch the actual config - just the local instance that the dialog has. It won't get modified until the OK button is pressed, and is discarded when the cancel button is pressed.